backend/requirements.txt
certifi==2018.1.18
chardet==3.0.4
cookies==2.2.1
decorator==4.2.1
Django==2.0.2
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
djangorestframework==3.7.7
googlemaps==2.5.1
haversine==0.4.5
idna==2.6
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
mock==2.0.0
parso==0.1.1
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyowm==2.8.0
pytz==2018.3
requests==2.18.4
responses==0.8.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
zipcode==2.0.0
